Refactor RGBA to RBG
