+++ src/color/RGBInput/component.js
