+++ src/color/RGBInput/index.js
