+++ src/color/RGBInput/styles.js
