+++ src/color/picker-area/component.js
