+++ src/color/picker-area/enhancer/connect.js
