+++ src/color/picker-area/index.js
