+++ src/color/picker-area/styles.js
